requirements.txt
fastapi>=0.85.0
uvicorn>=0.19.0
python-dotenv>=0.21.0
tensorflow-cpu>=2.10.0
gunicorn>=20.1.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
joblib>=1.2.0